Unify logic for laying out floats and atomic inlines
